requirements.txt
websocket-client==1.4.1
websockets==10.3
fastapi==0.85.0
uvicorn==0.18.3